Restructure code into files and fix translate extent
