+++ src/components/BottomBar/Missions/Timeline.scss
@@ -5,6 +5,7 @@ $font-size: 0.8rem;
 .bar {
   fill: grey;
   opacity: 0.9;
+  stroke: black;
 }
 
 .barHighlighted {
